requirements.txt
gurobipy
networkx
numpy
matplotlib
tqdm

# server
flask==3.0.3
flask-cors==4.0.0
psycopg2-binary